requirements.txt
deepface==0.0.93
edge_tts==7.0.2
numpy==2.3.1
Pillow==11.3.0
pygame==2.6.1
scipy==1.16.0
sounddevice==0.5.2
streamlit==1.45.1
torch==2.7.1
tqdm==4.67.1
transformers==4.53.2
SpeechRecognition==3.14.3
opencv-python==4.12.0.88
sentencepiece==0.2.0
tokenizers==0.21.2
huggingface-hub==0.33.4
requests==2.32.4
